\name{trec_fast_test-package}
\alias{trec_fast_test-package}
\alias{trec_fast_test}
\docType{package}
\title{
\packageTitle{trec_fast_test}
}
\description{
\packageDescription{trec_fast_test}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trec_fast_test}
\packageIndices{trec_fast_test}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{trec_fast_test}

Maintainer: \packageMaintainer{trec_fast_test}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
